+++ src/main/java/org/huangsam/sample/numerical/NumberReporter.java
@@ -6,7 +6,7 @@ import org.slf4j.LoggerFactory;
 public class NumberReporter {
     private static final Logger LOG = LoggerFactory.getLogger(NumberReporter.class);
 
-    public void report(Integer result, Integer id) {
+    public void report(Long result, Integer id) {
         LOG.info("Crunched {} with thread {}", result, id);
     }
 }
